Update welcome.yml
Workflows have jobs, and jobs have steps. So now we'll add a step to your workflow.

What are steps?: Actions steps run - in the order they are specified, from the top down - when a workflow job is processed. Each step must pass for the next step to run.

Each step consists of either a shell script that's executed, or a reference to an action that's run. When we talk about an action (with a lowercase "a") in this context, we mean a reusable unit of code. You can find out about actions in "Finding and customizing actions," but for now we'll use a shell script in our workflow step.

Update your workflow to make it post a comment on new pull requests. It will do this using a bash script and GitHub CLI.

⌨️ Activity: Add a step to your workflow file
Still working on the welcome-workflow branch, open your welcome.yml file.

Update the contents of the file to:
x
x
Note: The step you've added uses GitHub CLI (gh) to add a comment when a pull request is opened. To allow GitHub CLI to post a comment, we set the GITHUB_TOKEN environment variable to the value of the GITHUB_TOKEN secret, which is an installation access token, created when the workflow runs. For more information, see "Automatic token authentication." We set the PR_URL environment variable to the URL of the newly created pull request, and we use this in the gh command.

Click Commit changes in the top right of the workflow editor.

Type your commit message and commit your changes directly to your branch.