Update README.md
Activity: Trigger the workflow
Make a new branch named test-workflow.
Make a change, such as adding an emoji to your README.md file, and commit the change directly to your new branch.
In the Pull requests tab, create a pull request that will merge test-workflow into main.
Watch the workflow running in the checks section of the pull request.
Notice the comment that the workflow adds to the pull request.

The shell script in the workflow will run whenever a new pull request is opened.
Seeing your action in action: The status of each workflow run that's triggered is shown in the pull request before it's merged: look for All checks have passed when you try out the steps below. You can also see a list of all the workflows that are running, or have finished running, in the Actions tab of your repository. From there, you can click on each workflow run to view more details and access log files.